Add dynamic condition ratings list for treatments
Introduces `TreatmentConditionRatingsList` component and refactors the ratings page for better modularity. Centralized data fetching logic into a new `getTreatmentWithConditionRatings` server action, improving code maintainability and readability.
